Refactored TaxonomyTree into a separate component
