+++ src/controls/modernTaxonomyPicker/taxonomyPanelContents/TaxonomyPanelContents.tsx
@@ -1,65 +1,28 @@
 import * as React from 'react';
 import styles from './TaxonomyPanelContents.module.scss';
-import { Checkbox,
-         ChoiceGroup,
-         FocusZone,
-         FocusZoneDirection,
-         getRTLSafeKeyCode,
-         GroupedList,
-         GroupHeader,
-         IBasePickerStyleProps,
+import { IBasePickerStyleProps,
          IBasePickerStyles,
-         ICheckboxStyleProps,
-         ICheckboxStyles,
-         IChoiceGroupOption,
-         IChoiceGroupOptionStyleProps,
-         IChoiceGroupOptionStyles,
-         IChoiceGroupStyleProps,
-         IChoiceGroupStyles,
-         IGroup,
-         IGroupFooterProps,
-         IGroupHeaderProps,
-         IGroupHeaderStyleProps,
-         IGroupHeaderStyles,
-         IGroupRenderProps,
-         IGroupShowAllProps,
-         ILabelStyleProps,
-         ILabelStyles,
-         ILinkStyleProps,
-         ILinkStyles,
-         IListProps,
          IPickerItemProps,
-         IRenderFunction,
-         ISpinnerStyleProps,
-         ISpinnerStyles,
          IStyleFunctionOrObject,
          ISuggestionItemProps,
-         KeyCodes,
          Label,
-         Link,
          Selection,
-         SelectionMode,
-         SelectionZone,
-         Spinner
        } from 'office-ui-fabric-react';
 import { ITermInfo,
          ITermSetInfo,
          ITermStoreInfo
        } from '@pnp/sp/taxonomy';
 import { Guid } from '@microsoft/sp-core-library';
-import { BaseComponentContext } from '@microsoft/sp-component-base';
-import { css } from '@uifabric/utilities/lib/css';
 import * as strings from 'ControlStrings';
 import { useForceUpdate } from '@uifabric/react-hooks';
 import { ModernTermPicker } from '../modernTermPicker/ModernTermPicker';
 import { IReadonlyTheme } from "@microsoft/sp-component-base";
 import { IModernTermPickerProps } from '../modernTermPicker/ModernTermPicker.types';
 import { Optional } from '../ModernTaxonomyPicker';
+import { TaxonomyTree } from '../taxonomyTree/TaxonomyTree';
 
 export interface ITaxonomyPanelContentsProps {
-  context: BaseComponentContext;
   allowMultipleSelections?: boolean;
-  termSetId: Guid;
   pageSize: number;
   selectedPanelOptions: ITermInfo[];
   setSelectedPanelOptions: React.Dispatch<React.SetStateAction<ITermInfo[]>>;
@@ -79,8 +42,6 @@ export interface ITaxonomyPanelContentsProps {
 }
 
 export function TaxonomyPanelContents(props: ITaxonomyPanelContentsProps): React.ReactElement<ITaxonomyPanelContentsProps> {
-  const [groupsLoading, setGroupsLoading] = React.useState<string[]>([]);
-  const [groups, setGroups] = React.useState<IGroup[]>([]);
   const [terms, setTerms] = React.useState<ITermInfo[]>(props.selectedPanelOptions?.length > 0 ? [...props.selectedPanelOptions] : []);
 
   const forceUpdate = useForceUpdate();
@@ -101,355 +62,6 @@ export function TaxonomyPanelContents(props: ITaxonomyPanelContentsProps): React
     return s;
   }, [terms]);
 
-  React.useEffect(() => {
-    let termRootName = "";
-    if (props.anchorTermInfo) {
-      let anchorTermNames = props.anchorTermInfo.labels.filter((name) => name.languageTag === props.languageTag && name.isDefault);
-      if (anchorTermNames.length === 0) {
-        anchorTermNames = props.anchorTermInfo.labels.filter((name) => name.languageTag === props.termStoreInfo.defaultLanguageTag && name.isDefault);
-      }
-      termRootName = anchorTermNames[0].name;
-    }
-    else {
-      let termSetNames = props.termSetInfo.localizedNames.filter((name) => name.languageTag === props.languageTag);
-      if (termSetNames.length === 0) {
-        termSetNames = props.termSetInfo.localizedNames.filter((name) => name.languageTag === props.termStoreInfo.defaultLanguageTag);
-      }
-      termRootName = termSetNames[0].name;
-    }
-    const rootGroup: IGroup = {
-      name: termRootName,
-      key: props.anchorTermInfo ? props.anchorTermInfo.id : props.termSetInfo.id,
-      startIndex: -1,
-      count: 50,
-      level: 0,
-      isCollapsed: false,
-      data: { skiptoken: '' },
-      hasMoreData: (props.anchorTermInfo ? props.anchorTermInfo.childrenCount : props.termSetInfo.childrenCount) > 0
-    };
-    setGroups([rootGroup]);
-    setGroupsLoading((prevGroupsLoading) => [...prevGroupsLoading, props.termSetInfo.id]);
-    if (props.termSetInfo.childrenCount > 0) {
-      props.onLoadMoreData(props.termSetId, props.anchorTermInfo ? Guid.parse(props.anchorTermInfo.id) : Guid.empty, '', true)
-        .then((loadedTerms) => {
-          const grps: IGroup[] = loadedTerms.value.map(term => {
-            let termNames = term.labels.filter((termLabel) => (termLabel.languageTag === props.languageTag && termLabel.isDefault === true));
-            if (termNames.length === 0) {
-              termNames = term.labels.filter((termLabel) => (termLabel.languageTag === props.termStoreInfo.defaultLanguageTag && termLabel.isDefault === true));
-            }
-            const g: IGroup = {
-              name: termNames[0]?.name,
-              key: term.id,
-              startIndex: -1,
-              count: 50,
-              level: 1,
-              isCollapsed: true,
-              data: { skiptoken: '', term: term },
-              hasMoreData: term.childrenCount > 0,
-            };
-            if (g.hasMoreData) {
-              g.children = [];
-            }
-            return g;
-          });
-          setTerms((prevTerms) => {
-            const nonExistingTerms = loadedTerms.value.filter((term) => prevTerms.every((prevTerm) => prevTerm.id !== term.id));
-            return [...prevTerms, ...nonExistingTerms];
-          });
-          rootGroup.children = grps;
-          rootGroup.data.skiptoken = loadedTerms.skiptoken;
-          rootGroup.hasMoreData = loadedTerms.skiptoken !== '';
-          setGroupsLoading((prevGroupsLoading) => prevGroupsLoading.filter((value) => value !== props.termSetId.toString()));
-          setGroups([rootGroup]);
-        });
-    }
-  }, []);
-
-  const onToggleCollapse = (group: IGroup): void => {
-    if (group.isCollapsed === true) {
-      setGroups((prevGroups) => {
-        const recurseGroups = (currentGroup: IGroup) => {
-          if (currentGroup.key === group.key) {
-            currentGroup.isCollapsed = false;
-          }
-          if (currentGroup.children?.length > 0) {
-            for (const child of currentGroup.children) {
-              recurseGroups(child);
-            }
-          }
-        };
-        let newGroupsState: IGroup[] = [];
-        for (const prevGroup of prevGroups) {
-          recurseGroups(prevGroup);
-          newGroupsState.push(prevGroup);
-        }
-
-        return newGroupsState;
-      });
-
-      if (group.children && group.children.length === 0) {
-        setGroupsLoading((prevGroupsLoading) => [...prevGroupsLoading, group.key]);
-        group.data.isLoading = true;
-
-        props.onLoadMoreData(props.termSetId, Guid.parse(group.key), '', true)
-          .then((loadedTerms) => {
-            const grps: IGroup[] = loadedTerms.value.map(term => {
-              let termNames = term.labels.filter((termLabel) => (termLabel.languageTag === props.languageTag && termLabel.isDefault === true));
-              if (termNames.length === 0) {
-                termNames = term.labels.filter((termLabel) => (termLabel.languageTag === props.termStoreInfo.defaultLanguageTag && termLabel.isDefault === true));
-              }
-              const g: IGroup = {
-                name: termNames[0]?.name,
-                key: term.id,
-                startIndex: -1,
-                count: 50,
-                level: group.level + 1,
-                isCollapsed: true,
-                data: { skiptoken: '', term: term },
-                hasMoreData: term.childrenCount > 0,
-              };
-              if (g.hasMoreData) {
-                g.children = [];
-              }
-              return g;
-            });
-
-            setTerms((prevTerms) => {
-              const nonExistingTerms = loadedTerms.value.filter((term) => prevTerms.every((prevTerm) => prevTerm.id !== term.id));
-              return [...prevTerms, ...nonExistingTerms];
-            });
-
-            group.children = grps;
-            group.data.skiptoken = loadedTerms.skiptoken;
-            group.hasMoreData = loadedTerms.skiptoken !== '';
-            setGroupsLoading((prevGroupsLoading) => prevGroupsLoading.filter((value) => value !== group.key));
-          });
-      }
-    }
-    else {
-      setGroups((prevGroups) => {
-        const recurseGroups = (currentGroup: IGroup) => {
-          if (currentGroup.key === group.key) {
-            currentGroup.isCollapsed = true;
-          }
-          if (currentGroup.children?.length > 0) {
-            for (const child of currentGroup.children) {
-              recurseGroups(child);
-            }
-          }
-        };
-        let newGroupsState: IGroup[] = [];
-        for (const prevGroup of prevGroups) {
-          recurseGroups(prevGroup);
-          newGroupsState.push(prevGroup);
-        }
-
-        return newGroupsState;
-      });
-
-    }
-  };
-
-  const onRenderTitle = (groupHeaderProps: IGroupHeaderProps) => {
-    const isChildSelected = (children: IGroup[]): boolean => {
-      let aChildIsSelected = children && children.some((child) => selection.isKeySelected(child.key) || isChildSelected(child.children));
-      return aChildIsSelected;
-    };
-
-    const childIsSelected = isChildSelected(groupHeaderProps.group.children);
-
-    if (groupHeaderProps.group.level === 0) {
-      const labelStyles: IStyleFunctionOrObject<ILabelStyleProps, ILabelStyles> = {root: {width: "100%", fontWeight: childIsSelected ? "bold" : "normal"}};
-      return (
-        <FocusZone
-          direction={FocusZoneDirection.horizontal}
-          className={styles.taxonomyItemFocusZone}
-        >
-          <Label styles={labelStyles}>{groupHeaderProps.group.name}</Label>
-          <div className={styles.actionButtonContainer}>
-            {props.onRenderActionButton && props.onRenderActionButton(props.termStoreInfo, props.termSetInfo, props.anchorTermInfo)}
-          </div>
-        </FocusZone>
-      );
-    }
-
-    const isDisabled = groupHeaderProps.group.data.term.isAvailableForTagging.filter((t) => t.setId === props.termSetId.toString())[0].isAvailable === false;
-    const isSelected = selection.isKeySelected(groupHeaderProps.group.key);
-
-    if (props.allowMultipleSelections) {
-      const checkBoxStyles: IStyleFunctionOrObject<ICheckboxStyleProps, ICheckboxStyles> = {root: { flex: "1" } };
-      if (isSelected || childIsSelected) {
-        checkBoxStyles.label = { fontWeight: 'bold' };
-      }
-      else {
-        checkBoxStyles.label = { fontWeight: 'normal' };
-      }
-
-      return (
-        <FocusZone
-          direction={FocusZoneDirection.horizontal}
-          className={styles.taxonomyItemFocusZone}
-        >
-          <Checkbox
-            key={groupHeaderProps.group.key}
-            label={groupHeaderProps.group.name}
-            checked={isSelected}
-            styles={checkBoxStyles}
-            disabled={isDisabled}
-            onRenderLabel={(p) => <span className={css(!isDisabled && styles.checkbox, isDisabled && styles.disabledCheckbox, isSelected && styles.selectedCheckbox)} title={p.title}>
-              {p.label}
-            </span>}
-            onChange={(ev?: React.FormEvent<HTMLElement | HTMLInputElement>, checked?: boolean) => {
-              selection.setKeySelected(groupHeaderProps.group.key, checked, false);
-            }}
-          />
-          <div className={styles.actionButtonContainer}>
-            {props.onRenderActionButton && props.onRenderActionButton(props.termStoreInfo, props.termSetInfo, groupHeaderProps.group.data.term)}
-          </div>
-        </FocusZone>
-      );
-    }
-    else {
-      const choiceGroupOptionStyles: IStyleFunctionOrObject<IChoiceGroupOptionStyleProps, IChoiceGroupOptionStyles> = isSelected || childIsSelected ? { root: {marginTop: 0}, choiceFieldWrapper: { fontWeight: 'bold', flex: '1' }, field: { width: '100%'} } : { root: {marginTop: 0}, choiceFieldWrapper: { fontWeight: 'normal', flex: '1' }, field: { width: '100%'} };
-      const options: IChoiceGroupOption[] = [{
-                                                key: groupHeaderProps.group.key,
-                                                text: groupHeaderProps.group.name,
-                                                styles: choiceGroupOptionStyles,
-                                                onRenderLabel: (p) =>
-                                                  <span id={p.labelId} className={css(!isDisabled && styles.choiceOption, isDisabled && styles.disabledChoiceOption, isSelected && styles.selectedChoiceOption)}>
-                                                    {p.text}
-                                                  </span>,
-                                                onClick: () => {
-                                                  selection.setAllSelected(false);
-                                                  selection.setKeySelected(groupHeaderProps.group.key, true, false);
-                                                }
-                                              }];
-
-      const choiceGroupStyles: IStyleFunctionOrObject<IChoiceGroupStyleProps, IChoiceGroupStyles> = { root: { flex: "1" }, applicationRole: { width: "100%" } };
-
-      return (
-        <FocusZone
-          direction={FocusZoneDirection.horizontal}
-          className={styles.taxonomyItemFocusZone}
-        >
-            <ChoiceGroup
-              options={options}
-              selectedKey={selection.getSelection()[0]?.id}
-              disabled={isDisabled}
-              styles={choiceGroupStyles}
-            />
-          <div className={styles.actionButtonContainer}>
-            {props.onRenderActionButton && props.onRenderActionButton(props.termStoreInfo, props.termSetInfo, groupHeaderProps.group.data.term)}
-          </div>
-        </FocusZone>
-      );
-    }
-  };
-
-  const onRenderHeader = (headerProps: IGroupHeaderProps): JSX.Element => {
-    const groupHeaderStyles: IStyleFunctionOrObject<IGroupHeaderStyleProps, IGroupHeaderStyles> = {
-      expand: { height: 42, visibility: !headerProps.group.children || headerProps.group.level === 0 ? "hidden" : "visible", fontSize: 14 },
-      expandIsCollapsed: { visibility: !headerProps.group.children || headerProps.group.level === 0 ? "hidden" : "visible", fontSize: 14 },
-      check: { display: 'none' },
-      headerCount: { display: 'none' },
-      groupHeaderContainer: { height: 36, paddingTop: 3, paddingBottom: 3, paddingLeft: 3, paddingRight: 3, alignItems: 'center', },
-      root: { height: 42 },
-    };
-
-    const isDisabled = headerProps.group.data.term && headerProps.group.data.term.isAvailableForTagging.filter((t) => t.setId === props.termSetId.toString())[0].isAvailable === false;
-
-    return (
-      <GroupHeader
-        {...headerProps}
-        styles={groupHeaderStyles}
-        className={styles.taxonomyItemHeader}
-        onRenderTitle={onRenderTitle}
-        onToggleCollapse={onToggleCollapse}
-        indentWidth={20}
-        expandButtonProps={{style: {color: props.themeVariant?.semanticColors.bodyText}}}
-        onGroupHeaderKeyUp={(ev: React.KeyboardEvent<HTMLElement>, group: IGroup) => {
-          if ((ev.keyCode == 32 || ev.keyCode == 13) && !isDisabled) {
-            if (props.allowMultipleSelections) {
-              selection.toggleKeySelected(headerProps.group.key);
-            }
-            else {
-              selection.setAllSelected(false);
-              selection.setKeySelected(headerProps.group.key, true, false);
-            }
-          }
-        }}
-      />
-    );
-  };
-
-  const onRenderFooter = (footerProps: IGroupFooterProps): JSX.Element => {
-    if ((footerProps.group.hasMoreData || footerProps.group.children && footerProps.group.children.length === 0) && !footerProps.group.isCollapsed) {
-
-      if (groupsLoading.some(value => value === footerProps.group.key)) {
-        const spinnerStyles: IStyleFunctionOrObject<ISpinnerStyleProps, ISpinnerStyles> = { circle: { verticalAlign: 'middle' } };
-        return (
-          <div className={styles.spinnerContainer}>
-            <Spinner styles={spinnerStyles} />
-          </div>
-        );
-      }
-      const linkStyles: IStyleFunctionOrObject<ILinkStyleProps, ILinkStyles> = { root: { fontSize: '14px', paddingLeft: (footerProps.groupLevel + 1) * 20 + 62 } };
-      return (
-        <div className={styles.loadMoreContainer}>
-          <Link onClick={() => {
-            setGroupsLoading((prevGroupsLoading) => [...prevGroupsLoading, footerProps.group.key]);
-            props.onLoadMoreData(props.termSetId, footerProps.group.key === props.termSetId.toString() ? Guid.empty : Guid.parse(footerProps.group.key), footerProps.group.data.skiptoken, true)
-              .then((loadedTerms) => {
-                const grps: IGroup[] = loadedTerms.value.map(term => {
-                  let termNames = term.labels.filter((termLabel) => (termLabel.languageTag === props.languageTag && termLabel.isDefault === true));
-                  if (termNames.length === 0) {
-                    termNames = term.labels.filter((termLabel) => (termLabel.languageTag === props.termStoreInfo.defaultLanguageTag && termLabel.isDefault === true));
-                  }
-                  const g: IGroup = {
-                    name: termNames[0]?.name,
-                    key: term.id,
-                    startIndex: -1,
-                    count: 50,
-                    level: footerProps.group.level + 1,
-                    isCollapsed: true,
-                    data: { skiptoken: '', term: term },
-                    hasMoreData: term.childrenCount > 0,
-                  };
-                  if (g.hasMoreData) {
-                    g.children = [];
-                  }
-                  return g;
-                });
-                setTerms((prevTerms) => {
-                  const nonExistingTerms = loadedTerms.value.filter((term) => prevTerms.every((prevTerm) => prevTerm.id !== term.id));
-                  return [...prevTerms, ...nonExistingTerms];
-                });
-                footerProps.group.children = [...footerProps.group.children, ...grps];
-                footerProps.group.data.skiptoken = loadedTerms.skiptoken;
-                footerProps.group.hasMoreData = loadedTerms.skiptoken !== '';
-                setGroupsLoading((prevGroupsLoading) => prevGroupsLoading.filter((value) => value !== footerProps.group.key));
-              });
-          }}
-            styles={linkStyles}>
-            {strings.ModernTaxonomyPickerLoadMoreText}
-          </Link>
-        </div>
-      );
-    }
-    return null;
-  };
-
-  const onRenderShowAll: IRenderFunction<IGroupShowAllProps> = () => {
-    return null;
-  };
-
-  const groupProps: IGroupRenderProps = {
-    onRenderFooter: onRenderFooter,
-    onRenderHeader: onRenderHeader,
-    showEmptyGroups: true,
-    onRenderShowAll: onRenderShowAll,
-  };
-
   const onPickerChange = (items?: ITermInfo[]): void => {
     const itemsToAdd = items.filter((item) => terms.every((term) => term.id !== item.id));
     setTerms((prevTerms) => [...prevTerms, ...itemsToAdd]);
@@ -459,10 +71,6 @@ export function TaxonomyPanelContents(props: ITaxonomyPanelContentsProps): React
         selection.setKeySelected(item.id.toString(), true, false);
       }
     }
-  };
-
-  const shouldEnterInnerZone = (ev: React.KeyboardEvent<HTMLElement>): boolean => {
-    return ev.which === getRTLSafeKeyCode(KeyCodes.right);
   };
 
   const termPickerStyles: IStyleFunctionOrObject<IBasePickerStyleProps, IBasePickerStyles> = { root: {paddingTop: 4, paddingBottom: 4, paddingRight: 4, minheight: 34}, input: {minheight: 34}, text: { minheight: 34, borderStyle: 'none', borderWidth: '0px' } };
@@ -492,17 +100,19 @@ export function TaxonomyPanelContents(props: ITaxonomyPanelContentsProps): React
         </div>
       </div>
       <Label className={styles.taxonomyTreeLabel}>{strings.ModernTaxonomyPickerTreeTitle}</Label>
-      <div>
-        <GroupedList
-          items={[]}
-          onRenderCell={null}
-          groups={groups}
-          groupProps={groupProps}
-          onShouldVirtualize={(p: IListProps<any>) => false}
-          data-is-focusable={true}
-          focusZoneProps={{direction: FocusZoneDirection.vertical, shouldEnterInnerZone: shouldEnterInnerZone}}
-        />
-      </div>
+      <TaxonomyTree
+        anchorTermInfo={props.anchorTermInfo}
+        languageTag={props.languageTag}
+        onLoadMoreData={props.onLoadMoreData}
+        pageSize={props.pageSize}
+        selection={selection}
+        setTerms={setTerms}
+        termSetInfo={props.termSetInfo}
+        termStoreInfo={props.termStoreInfo}
+        terms={terms}
+        allowMultipleSelections={props.allowMultipleSelections}
+        onRenderActionButton={props.onRenderActionButton}
+      />
     </div>
   );
 }
